requirements.txt
# AI Trading System - Requirements (Python 3.12.9 compatible)
# Core data stack
pandas==2.2.0
numpy==1.26.0

# Data collection
vnstock==3.4.2
requests==2.31.0
beautifulsoup4==4.12.2

# Database
sqlalchemy==2.0.25

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.1
schedule==1.2.1

# Dev tools
jupyter==1.0.0
ipython==8.22.2
black==24.1.1
pytest==7.4.4

# Python 3.12 compatibility
typing-extensions==4.10.0
